fastapi/requirements.txt
pymongo
dnspython
ipaddress
uvicorn
fastapi
motor